feat: integrate node-oidc-provider by panva
This commit is responsible for integrating the node-oidc-provider created by panva and
removing the custom implementation of oidc. As a result there was a change in schema and
a lot of things got deprecated and broken. Tests broke, code failed. Deprecated code was removed and All of the mess was
cleaned and stabilized. Zolo.md is created to help understand the next developer the
remaining tasks and how to accomplish them. Tests were removed as a result the count came
down to 386 from 510 :(