feat: move to node-oidc-provider (#82)
* feat: integrate node-oidc-provider by panva

This commit is responsible for integrating the node-oidc-provider created by panva and
removing the custom implementation of oidc. As a result there was a change in schema and
a lot of things got deprecated and broken. Tests broke, code failed. Deprecated code was removed and All of the mess was
cleaned and stabilized. Zolo.md is created to help understand the next developer the
remaining tasks and how to accomplish them. Tests were removed as a result the count came
down to 386 from 510 :(

* chore: update kickstart.json

* fix: integrate sqlite

* refactor: update code to remove errors

This commit is responsible for removing server_error from the devInteractions page

* fix: start using nest oidc provider

* refactor: make nest-oidc-provider to handle login

Nest oidc provider was used to made login process smooth and easy. Some errors were
encountered since sqlite was used. however was corrected

* fix: user can login with correct password

Earlier a user/someone else can login from any password value, however it is changed to ->
match the original password value to avoid unauthorized access

* fix: minio policy login issue

Minio requires that the AS supports a policy claim in the id_token, hence it needed to be mapped
using the roles available to a user in that tenant, now the user is login based on admin or normal user

* feat: add consentless login

Now applications can login without asking for consent from the end user

This feature has been warned not to used because 1. asking for user consent is ethical
2. This might be not be bug-free as mentioned below

link: https://github.com/panva/node-oidc-provider/blob/v7.x/recipes/skip_consent.md

* fix: prevent possible privilege escalation

earlier findAccount just returns the role and don't check whether the role belongs to
corresponding application or not which might have led to privilege escalation. Now it
is fixed in the findAccount function

* test: update tests

This commit fixes failing tests due to change in schema of application and other minor fixes

---------

Co-authored-by: Zolo-Ryan <utkarshswaroopsri187@gmail.com>
Co-authored-by: Zolo Ryan <119104324+Zolo-Ryan@users.noreply.github.com>